Split code files #26
